Sanitizer: Add in build environment support
configure --enable-sanitizer
cmake -DENABLE_SANITIZER=ON

Addresses trivial issues raised by the ASan sanitizer.

Includes m4/ax_check_compile_flag.m4, which is a part of the
pkg autoconf-archive which not all systems have.

+++ ccm.c
@@ -1,6 +1,6 @@
 /*******************************************************************************
  *
- * Copyright (c) 2011, 2012, 2013, 2014, 2015 Olaf Bergmann (TZI) and others.
+ * Copyright (c) 2011-2026 Olaf Bergmann (TZI) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * and Eclipse Distribution License v. 1.0 which accompanies this distribution.
@@ -28,7 +28,7 @@
 
 #define CCM_FLAGS(A,M,L) (((A > 0) << 6) | (((M - 2)/2) << 3) | (L - 1))
 
-#define MASK_L(_L) ((1 << 8 * _L) - 1)
+#define MASK_L(_L) (((uint64_t)1 << (8 * _L)) - 1)
 
 #define SET_COUNTER(A,L,cnt,C) {					\
     unsigned int i_;                                                    \
